An attempt at a 1.19 update (broken)
Works as long as you don't change any settings while in game (aka do it in the file)

Why? : When the config is changed ingame, MixinWorldRenderer.reloadStars() is called. This does something with BufferBuilder where no matter what, it makes the game crash with "java.lang.IllegalStateException: Already building!", even though the buffer is in fact NOT building (as evidenced by my println()s). For the time being, I have given up on solving this because I no longer know what I'm doing (not that I did in the first place). For (slightly) more info, see this message in the FabricMC discord: https://discord.com/channels/507304429255393322/507982478276034570/984315799571410964